Invite Gemini CLI to the party.
